+++ packages/client/src/index.ts
@@ -10,6 +10,7 @@
  */
 
 export * from './workflow-client';
+export * from './async-completion-client';
 export * from './connection';
 export * from './types';
 export * from './errors';
@@ -27,4 +28,5 @@ export {
   TerminatedFailure,
   TimeoutFailure,
   ProtoFailure,
+  RetryPolicy,
 } from '@temporalio/common';