+++ packages/common/src/workflow-options.ts
@@ -1,10 +1,19 @@
-import { coresdk, google } from '@temporalio/proto/lib/coresdk';
+import type { coresdk, google } from '@temporalio/proto/lib/coresdk';
 import { Workflow } from './interfaces';
+import { RetryPolicy } from './retry-policy';
 import { msToTs } from './time';
+import { checkExtends } from './type-helpers';
 
-export type WorkflowIdReusePolicy = coresdk.common.WorkflowIdReusePolicy;
-export const WorkflowIdReusePolicy = coresdk.common.WorkflowIdReusePolicy;
-export type RetryPolicy = coresdk.common.IRetryPolicy;
+// Avoid importing the proto implementation to reduce workflow bundle size
+// Copied from coresdk.common.WorkflowIdReusePolicy
+export enum WorkflowIdReusePolicy {
+  WORKFLOW_ID_REUSE_POLICY_UNSPECIFIED = 0,
+  WORKFLOW_ID_REUSE_POLICY_ALLOW_DUPLICATE = 1,
+  WORKFLOW_ID_REUSE_POLICY_ALLOW_DUPLICATE_FAILED_ONLY = 2,
+  WORKFLOW_ID_REUSE_POLICY_REJECT_DUPLICATE = 3,
+}
+
+checkExtends<coresdk.common.WorkflowIdReusePolicy, WorkflowIdReusePolicy>();
 
 export interface BaseWorkflowOptions {
   /**
@@ -18,7 +27,13 @@ export interface BaseWorkflowOptions {
    */
   workflowIdReusePolicy?: WorkflowIdReusePolicy;
 
-  retryPolicy?: coresdk.common.IRetryPolicy;
+  /**
+   * Controls how a Workflow is retried.
+   *
+   * Workflows should typically use the system default, do not set this unless
+   * you know what you're doing.
+   */
+  retry?: RetryPolicy;
 
   /**
    * Optional cron schedule for Workflow. If a cron schedule is specified, the Workflow will run
@@ -98,12 +113,11 @@ export type WithCompiledWorkflowDurationOptions<T extends WorkflowDurationOption
   workflowTaskTimeout?: google.protobuf.IDuration;
 };
 
-export function compileWorkflowOptions<T extends WorkflowDurationOptions>({
-  workflowExecutionTimeout,
-  workflowRunTimeout,
-  workflowTaskTimeout,
-  ...rest
-}: T): WithCompiledWorkflowDurationOptions<T> {
+export function compileWorkflowOptions<T extends WorkflowDurationOptions>(
+  options: T
+): WithCompiledWorkflowDurationOptions<T> {
+  const { workflowExecutionTimeout, workflowRunTimeout, workflowTaskTimeout, ...rest } = options;
+
   return {
     ...rest,
     workflowExecutionTimeout: workflowExecutionTimeout ? msToTs(workflowExecutionTimeout) : undefined,