+++ packages/test/src/workflows/signal-target.ts
@@ -5,11 +5,17 @@
  * @module
  */
 import { condition, setHandler } from '@temporalio/workflow';
-import { failWithMessageSignal, unblockSignal } from './definitions';
+import { argsTestSignal, failWithMessageSignal, unblockSignal } from './definitions';
 
 export async function signalTarget(): Promise<void> {
   let unblocked = false;
 
+  // Verify arguments are sent correctly
+  setHandler(argsTestSignal, (num, str) => {
+    if (!(num === 123 && str === 'kid')) {
+      throw new Error('Invalid arguments');
+    }
+  });
   setHandler(failWithMessageSignal, (message) => {
     throw new Error(message);
   });