+++ packages/worker/src/interceptors.ts
@@ -48,6 +48,8 @@ export interface WorkerInterceptors {
    * List of modules to search for Workflow interceptors in
    * - Modules should export an `interceptors` variable of type {@link WorkflowInterceptorsFactory}
    * - Workflow interceptors run in in the Workflow isolate
+   *
+   * **NOTE**: This options is not used if worker is provided with pre-built bundle.
    */
   workflowModules?: string[];
 }